Standardize item status functionality on the src side.
