+++ src/java/CdbWebPortal/src/java/gov/anl/aps/cdb/portal/controllers/ItemDomainInventoryBaseController.java
@@ -7,118 +7,68 @@ package gov.anl.aps.cdb.portal.controllers;
 import gov.anl.aps.cdb.portal.controllers.settings.ItemSettings;
 import gov.anl.aps.cdb.portal.model.db.beans.ItemFacadeBase;
 import gov.anl.aps.cdb.portal.model.db.beans.PropertyTypeFacade;
-import gov.anl.aps.cdb.portal.model.db.entities.AllowedPropertyValue;
-import gov.anl.aps.cdb.portal.model.db.entities.Domain;
 import gov.anl.aps.cdb.portal.model.db.entities.ItemDomainInventoryBase;
+import gov.anl.aps.cdb.portal.model.db.entities.LocatableStatusItem;
 import gov.anl.aps.cdb.portal.model.db.entities.PropertyType;
-import gov.anl.aps.cdb.portal.model.db.entities.PropertyTypeCategory;
 import gov.anl.aps.cdb.portal.model.db.entities.PropertyValue;
-import gov.anl.aps.cdb.portal.view.objects.InventoryStatusPropertyAllowedValue;
+import gov.anl.aps.cdb.portal.model.db.utilities.ItemStatusUtility;
 import gov.anl.aps.cdb.portal.view.objects.InventoryStatusPropertyTypeInfo;
-import gov.anl.aps.cdb.portal.view.objects.ItemCoreMetadataPropertyInfo;
-import java.util.ArrayList;
-import java.util.List;
 import javax.ejb.EJB;
 
 /**
  *
  * @author craig
  */
-public abstract class ItemDomainInventoryBaseController<ItemInventoryBaseDomainEntity extends ItemDomainInventoryBase, ItemDomainInventoryEntityBaseFacade extends ItemFacadeBase<ItemInventoryBaseDomainEntity>, ItemInventoryEntityBaseSettingsObject extends ItemSettings> extends ItemController<ItemInventoryBaseDomainEntity, ItemDomainInventoryEntityBaseFacade, ItemInventoryEntityBaseSettingsObject>  {
-    
+public abstract class ItemDomainInventoryBaseController<ItemInventoryBaseDomainEntity extends ItemDomainInventoryBase, ItemDomainInventoryEntityBaseFacade extends ItemFacadeBase<ItemInventoryBaseDomainEntity>, ItemInventoryEntityBaseSettingsObject extends ItemSettings> extends ItemController<ItemInventoryBaseDomainEntity, ItemDomainInventoryEntityBaseFacade, ItemInventoryEntityBaseSettingsObject> implements IItemStatusController {
+
     // Inventory status variables
     protected InventoryStatusPropertyTypeInfo inventoryStatusPropertyTypeInfo = null;
-    private PropertyType inventoryStatusPropertyType; 
+    private PropertyType inventoryStatusPropertyType;
 
     @EJB
     private PropertyTypeFacade propertyTypeFacade;
-    
-    protected abstract String getStatusPropertyTypeName();
 
     @Override
     protected void loadEJBResourcesManually() {
-        super.loadEJBResourcesManually(); 
+        super.loadEJBResourcesManually();
         propertyTypeFacade = PropertyTypeFacade.getInstance();
     }
-    
+
     // <editor-fold defaultstate="collapsed" desc="Inventory status implementation">
-    
-    protected abstract InventoryStatusPropertyTypeInfo initializeInventoryStatusPropertyTypeInfo();
-    
+
     public InventoryStatusPropertyTypeInfo getInventoryStatusPropertyTypeInfo() {
-        if (inventoryStatusPropertyTypeInfo == null) {
-            inventoryStatusPropertyTypeInfo = initializeInventoryStatusPropertyTypeInfo();
-        }
+        inventoryStatusPropertyTypeInfo = ItemStatusUtility.getInventoryStatusPropertyTypeInfo(this, inventoryStatusPropertyTypeInfo);
         return inventoryStatusPropertyTypeInfo;
     }
-    
-    private PropertyType prepareInventoryStatusPropertyType() {
-        InventoryStatusPropertyTypeInfo propInfo = getInventoryStatusPropertyTypeInfo();
-        PropertyTypeController propertyTypeController = PropertyTypeController.getInstance();
-        
-        PropertyType propertyType = propertyTypeController.createEntityInstance();
-        propertyType.setIsInternal(true);
-        propertyType.setName(getStatusPropertyTypeName());
-        
-        PropertyTypeCategory category = PropertyTypeCategoryController.getInstance().findByName("Status");
-        if (category != null) {
-            propertyType.setPropertyTypeCategory(category);
-        }
-        
-        List<Domain> allowedDomainList = new ArrayList<>();
-        allowedDomainList.add(getDefaultDomain());
-        propertyType.setAllowedDomainList(allowedDomainList);
-        
-        List<AllowedPropertyValue> apvList = new ArrayList<>();
-        for (InventoryStatusPropertyAllowedValue valInfo : propInfo.getValues()) {
-            AllowedPropertyValue apv = new AllowedPropertyValue();
-            apv.setValue(valInfo.getValue());
-            apv.setSortOrder(valInfo.getSortOrder());
-            apv.setPropertyType(propertyType);
-            apvList.add(apv);
-        }
-        propertyType.setAllowedPropertyValueList(apvList);
-        
-        propertyType.setDefaultValue(propInfo.getDefaultValue());
 
-        propertyTypeController.setCurrent(propertyType);
-        propertyTypeController.create(true, false); 
-        return propertyType; 
-    }
-        
     public PropertyType getInventoryStatusPropertyType() {
-        if (inventoryStatusPropertyType == null) {
-            inventoryStatusPropertyType = 
-                    propertyTypeFacade.findByName(getStatusPropertyTypeName()); 
-        }
-        if (inventoryStatusPropertyType == null) {
-            inventoryStatusPropertyType = prepareInventoryStatusPropertyType();
-        }
+        inventoryStatusPropertyType = ItemStatusUtility.getInventoryStatusPropertyType(this, propertyTypeFacade, inventoryStatusPropertyType);
         return inventoryStatusPropertyType;
-    }        
+    }
 
     public PropertyValue getCurrentStatusPropertyValue() {
-        return getCurrent().getInventoryStatusPropertyValue();
+        return ItemStatusUtility.getCurrentStatusPropertyValue(this);
     }
     
+    public PropertyValue getItemStatusPropertyValue(LocatableStatusItem item) {
+        return ItemStatusUtility.getItemStatusPropertyValue(item); 
+    }
+
     public void prepareEditInventoryStatus() {
-        prepareEditInventoryStatus(getCurrent());
+        ItemStatusUtility.prepareEditInventoryStatus(this);
     }
-    
-    public void prepareEditInventoryStatus(ItemInventoryBaseDomainEntity item) {
-        if (getCurrentStatusPropertyValue() == null) {
-            PropertyValue preparePropertyTypeValueAdd = preparePropertyTypeValueAdd(getInventoryStatusPropertyType()); 
-            item.setInventoryStatusPropertyValue(preparePropertyTypeValueAdd);
-        }
+
+    public void prepareEditInventoryStatus(LocatableStatusItem item) {
+        ItemStatusUtility.prepareEditInventoryStatus(this, item);
     }
-    
+
     public synchronized void prepareEditInventoryStatusFromApi(ItemInventoryBaseDomainEntity item) {
         setCurrent(item);
         prepareEditInventoryStatus();
     }
-    
+
     public boolean getRenderedHistoryButton() {
-        return getCurrentStatusPropertyValue() != null; 
+        return ItemStatusUtility.getRenderedHistoryButton(this);
     }
 
     // </editor-fold>
