+++ src/java/CdbWebPortal/src/java/gov/anl/aps/cdb/portal/model/db/entities/ItemDomainInventoryBase.java
@@ -10,15 +10,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
  *
  * @author djarosz
  */
-public abstract class ItemDomainInventoryBase <CatalogItemType extends ItemDomainCatalogBase> extends LocatableItem {
-    
-    // Inventory status variables
-    protected transient PropertyValue inventoryStatusPropertyValue;
-    protected transient boolean loadedCurrentStatusPropertyValue = false;
-
-    protected transient Boolean sparePartIndicator = null;
-    
-    public abstract String getStatusPropertyTypeName();
+public abstract class ItemDomainInventoryBase <CatalogItemType extends ItemDomainCatalogBase> extends LocatableStatusItem {
 
     public CatalogItemType getCatalogItem() {
         return (CatalogItemType) getDerivedFromItem();
@@ -27,44 +19,4 @@ public abstract class ItemDomainInventoryBase <CatalogItemType extends ItemDomai
     public void setCatalogItem(CatalogItemType catalogItem) {
         setDerivedFromItem(catalogItem);
     }
-
-    @JsonIgnore
-    public PropertyValue getInventoryStatusPropertyValue() {
-        if (!loadedCurrentStatusPropertyValue) {
-            if (this.getPropertyValueInternalList() != null) {
-                for (PropertyValue propertyValue : this.getPropertyValueInternalList()) {
-                    String propertyTypeName = propertyValue.getPropertyType().getName();
-                    if (propertyTypeName.equals(getStatusPropertyTypeName())) {
-                        inventoryStatusPropertyValue = propertyValue;
-                        break;
-                    }
-                }
-            }
-            loadedCurrentStatusPropertyValue = true;
-        }
-        return inventoryStatusPropertyValue;
-    }
-
-    public void setInventoryStatusPropertyValue(PropertyValue inventoryStatusPropertyValue) {
-        this.inventoryStatusPropertyValue = inventoryStatusPropertyValue;
-    }
-
-    @JsonIgnore
-    public String getInventoryStatusValue() {
-        if (getInventoryStatusPropertyValue() != null) {
-            String value = getInventoryStatusPropertyValue().getValue();
-            if (value != null) {
-                return value;
-            }
-        }
-        return "";
-    }
-
-    public void setInventoryStatusValue(String status) {
-        if (getInventoryStatusPropertyValue() != null) {
-            getInventoryStatusPropertyValue().setValue(status);
-            sparePartIndicator = null;
-        }
-    }
-
 }
